refactor: extract shared test mocks to testhelpers_test.go
